requirements.txt
Django==3.2.16
django-extensions==3.2.1
python-decouple==3.6
django-redis==5.2.0
psycopg2==2.9.3
shortuuid==1.0.9
django-simple-captcha==0.5.17
django-tailwind==3.4.0
django-ipware==4.0.2